Move some job-related stuff into `job.rs`
